Renamed AppContainer.
