+++ src/_containers/AppContainer.jsx
@@ -1,5 +1,5 @@
 import React, { useEffect } from 'react'
-import { setFavicon } from '../../_helpers/favicons'
+import { setFavicon } from '../_helpers/favicons'
 import { useLocation } from 'react-router-dom'
 
 export default function AppContainer(props) {
